requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
einops>=0.8.0
fvcore>=0.1.5
numpy>=1.24.0
nvidia-ml-py>=12.0.0
pandas>=2.0.0
plotly>=5.0.0
pynvml>=11.0.0
pytorch-metric-learning>=2.0.0
scikit-learn>=1.3.0
seaborn>=0.13.0
thop
timm>=1.0.0
torch==2.3.1
torchinfo>=1.8.0
torchmetrics>=1.0.0
torchsummary>=1.5.0
ultralytics-thop
matplotlib>=3.5.0
pyyaml
tensorboard
